moved base64 to own file
